Refactor priority tests to own file
